requirements.txt
py_midicsv
GitPython